Added support for AIX and tested it on version 7.1!
Added $tsm::rootgroup and changed config.pp, because in a default AIX setup the "root" group is not available.
Added $tsm::package_provider and $tsm::package_source to make the TSM package intstallable via NIM which is set as default in this module.

Found a bug in manifests/service/debian.pp (and maybe other). Subscription in the service type should point to Concat[$::tsm::config] instead of File[$::tsm::config] in order to make the service restart after a configuration change. Patched it in manifests/service/aix.pp.

Added support for AIX and tested it on version 7.1!
Added $tsm::rootgroup and changed config.pp, because in a default AIX setup the "root" group is not available.
Added $tsm::package_provider and $tsm::package_source to make the TSM package intstallable via NIM which is set as default in this module.

Found a bug in manifests/service/debian.pp (and maybe other). Subscription in the service type should point to Concat[$::tsm::config] instead of File[$::tsm::config] in order to make the service restart after a configuration change. Patched it in manifests/service/aix.pp.

wrote rspecs for aix support
set the variables provider and source to undef in order to not break the test