Make it possible to define multiple server stanzas
This change makes it possible to define multiple server stanzas via the
"tsm::stanzas" hash. The change does not break compatibility, so the
configuration of a single server with tsm class parameters is still
possible.

The dms.sys template is now splitted up further into a header, a section
for global setting and template for the tsm::config::stanza instances.